[Change] lib\filters.pyの作成とclean_query_paramsの共通化
